requirements.txt
asgiref==3.4.1
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.4.1
Django==3.2.6

django-cors-headers==3.7.0
django-filter==2.4.0
django-qrcode==0.3
djangorestframework==3.12.4
fonttools==4.27.1
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydyf==0.1.1
pyphen==0.11.0
pytz==2021.1
qrcode==7.3
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
webencodings==0.5.1
xlwt==1.3.0
zopfli==0.1.8
